implement segment deoverlapping
Previously, the algorithm calculated shadow mesh in the following
fashion:

1. Take all segments created by the collision mesh,
2. Filter out segments which do not face the player,
3. Extrude the line segments away from the player to make quads of
   shadow.

Unfortunately, quads created by the third step could overlap with each
other, which was a problem because they were rendered with transparency,
and the overlapping region was darker than the non-overlapping one.

This commit fixes this problem by introducing a step which modifies the
line segments so that they cover the same area, but do not overlap with
each other.

The algorithm very rarely produces frames with incorrect shadow, but is
good enough to work 99.9% of the time.